DASH-1009 Refactor utils, constants and helpers. Sorted by different logic types.
