+++ controller/constants/cron-jobs.js
@@ -1,0 +1,14 @@
+export const ORACLE_CACHE_KEYS = {
+  isBurnNFTOnPolygonJobExecuting: 'isBurnNFTOnPolygonJobExecuting',
+  isUnprocessedWrapActionsExecuting: 'isUnprocessedWrapActionsExecuting',
+  isWrapOnEthJobExecuting: 'isWrapOnEthJobExecuting',
+  isWrapOnPolygonJobExecuting: 'isWrapOnPolygonJobExecuting',
+  isUnwrapOnEthJobExecuting: 'isUnwrapOnEthJobExecuting',
+  isUnwrapOnPolygonJobExecuting: 'isUnwrapOnPolygonJobExecuting',
+  isUnprocessedUnwrapActionsOnPolygonExecuting:
+    'isUnprocessedUnwrapActionsOnPolygonExecuting',
+  isUnprocessedUnwrapActionsOnEthJobExecuting:
+    'isUnprocessedUnwrapActionsOnEthJobExecuting',
+  isUnprocessedBurnNFTActionsJobExecuting:
+    'isUnprocessedBurnNFTActionsJobExecuting',
+};
